@@ -1,10 +1,12 @@
 package ru.kmorozov.gbd.core.logic.Proxy.web;
 
+import com.google.api.client.util.Strings;
+import org.apache.commons.lang3.StringUtils;
 import org.jsoup.nodes.Document;
-import org.jsoup.nodes.Element;
-import org.jsoup.nodes.TextNode;
 
+import java.util.Arrays;
 import java.util.List;
+import java.util.regex.Pattern;
 import java.util.stream.Collectors;
 
 /**
@@ -12,7 +14,10 @@ import java.util.stream.Collectors;
  */
 public class SslProxiesListProvider extends AbstractProxyExtractor {
 
-    private static final String PROXY_LIST_URL = "http://www.sslproxies.org/";
+    private static final String PROXY_LIST_URL = "http://www.proxz.com/proxy_list_anonymous_us_0.html";
+    private static final String ipPortPattern = "^(([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5])\\.){3}([0-9]|[1-9][0-9]|1[0-9]{2}|2[0-4][0-9]|25[0-5]):[0-9]+";
+    private static final Pattern pattern = Pattern.compile(ipPortPattern, Pattern.LITERAL);
+    private static boolean checkRegexp = false;
 
     @Override
     protected String getProxyListUrl() {
@@ -21,10 +26,20 @@ public class SslProxiesListProvider extends AbstractProxyExtractor {
 
     @Override
     protected List<String> extractProxyList(final Document doc) {
-        return doc.getElementById("proxylisttable").getElementsByTag("tbody").get(0).getElementsByTag("tr").stream().map(SslProxiesListProvider::extractProxyData).collect(Collectors.toList());
+        String textWithProxies = doc.html().replaceAll("<", "|").replaceAll(">", "|");
+        return Arrays.stream(textWithProxies.split("\\|")).filter(s -> validIpPort(s)).collect(Collectors.toList());
     }
 
-    private static String extractProxyData(final Element element) {
-        return ((TextNode) element.childNode(0).childNode(0)).getWholeText() + ':' + ((TextNode) element.childNode(1).childNode(0)).getWholeText();
+    private boolean validIpPort(String str) {
+        if (Strings.isNullOrEmpty(str))
+            return false;
+
+        if (str.length() < 10 || str.length() > 20)
+            return false;
+
+        if (StringUtils.countMatches(str, ".") != 3 || !str.contains(":"))
+            return false;
+
+        return checkRegexp ? pattern.matcher(str).matches() : true;
     }
 }
